+++ rocketgram/api/edit_chat_invite_link.py
@@ -20,5 +20,7 @@ class EditChatInviteLink(ChatInviteLinkResultMixin, Request):
 
     chat_id: Union[int, str]
     invite_link: str
+    name: Optional[str] = None
     expire_date: Optional[datetime] = None
     member_limit: Optional[int] = None
+    creates_join_request: Optional[bool] = None